requirements.txt
Django==3.1
djangorestframework==3.12
drf-yasg==1.20.0